fix: implement bulk transfer dialog for asset management
